chore(enum): Create global enum file
