+++ src/components/CreateAdvertisement/SectionB.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { computed, ref, watch, onMounted } from 'vue'
 import { useI18n } from 'vue-i18n';
+import { Condition } from '@/enums/enums.ts'
 
 onMounted(() => {
   fetchCategories();
@@ -17,13 +18,6 @@ const emit = defineEmits<{
   (e: 'update:tags', value: string[]): void;
 }>()
 
-enum Condition {
-  New = 'NEW',
-  LikeNew = 'LIKE_NEW',
-  Good = 'GOOD',
-  Acceptable = 'ACCEPTABLE',
-  ForParts = 'FOR_PARTS'
-}
 
 
 const title = ref<string>('')
