+++ src/enums/enums.ts
@@ -1,0 +1,14 @@
+export enum Condition {
+  New = 'NEW',
+  LikeNew = 'LIKE_NEW',
+  Good = 'GOOD',
+  Acceptable = 'ACCEPTABLE',
+  ForParts = 'FOR_PARTS'
+}
+
+export enum PaymentMethod {
+  Direct = 'DIRECT',
+  Auction = 'AUCTION',
+  None = 'NONE'
+}
+
